Added BonusProperty
Added BonusProperty to hold BonusComponentValues (BaseBonus, Ability, Item etc)
Inlined m_charstat into (Get-/Change-)BaseStat methods
Moved m_buffMult1Bonus to Boni
Removed null check in StatCalculator.CalcValueFromBuffs (can never be null)
Removed null check in StatCalculator.CalcValueFromItems (can never be null)
Inlined StatCalculator static methods
Replaced all calls to GameLiving.<Category>Indexer[eProperty] with Boni.GetProperty(eProperty).<Category>
Removed comment in StatCalculator about BuffBonusCategory; doesn't apply anymore
Several clean-ups in UnitTests

+++ UnitTests/UT_GamePlayer.cs
@@ -122,7 +122,7 @@ namespace DOL.UnitTests.GameServer
         {
             var player = createPlayer(new CharacterClassAnimist());
             player.Level = 50;
-            player.BaseBuffBonusCategory[(int)eProperty.Acuity] = 50;
+            player.BaseBuffBonusCategory[eProperty.Acuity] = 50;
 
             int actual = player.Intelligence;
